before conclusion

+++ controllers/videoController.js
@@ -1,6 +1,9 @@
 import routes from "../routes";
 import Video from "../models/Video";
+import Comment from "../models/Comment";
 import { set } from "mongoose";
+
+
 
 export const home = async (req, res) => {
     try {
@@ -34,9 +37,11 @@ export const postUpload = async(req, res) => {
     const newVideo = await Video.create({
         fileUrl: path,
         title,
-        description
+        description,
+        creator: req.user.id
     });
-    console.log(newVideo);
+    req.user.videos.push(newVideo.id);
+    req.user.save();
     res.redirect(routes.videoDetail(newVideo.id));
 };
 export const videoDetail = async (req, res) => {
@@ -44,8 +49,10 @@ export const videoDetail = async (req, res) => {
         params: {id}
     } = req;
     try {
-        const video = await Video.findById(id);
-        res.render("videoDetail", { pageTitle: video.title, video });
+        const video = await Video.findById(id)
+            .populate('creator')
+            .populate('comments');
+        res.render("videoDetail", { pageTitle: video.title, video});
     } catch(error) {
         res.redirect(routes.home);
     }
@@ -56,11 +63,14 @@ export const getEditVideo = async (req, res) => {
     } = req;
     try {
         const video = await Video.findById(id);
-        res.render("editVideo", { pageTitle: `::EDIT::${video.title}`, video });
+        if(video.creator.toString() !== req.user.id) {
+            throw Error();
+        } else {
+            res.render("editVideo", { pageTitle: `::EDIT::${video.title}`, video });
+        }
     } catch(error) {
         res.redirect(routes.home);
     }
-    res.render("editVideo", { pageTitle: "::EDITVIDEO::" });
 }
 
 export const postEditVideo = async (req, res) => {
@@ -81,9 +91,52 @@ export const deleteVideo = async (req, res) => {
         params: {id}
     } = req;
     try {
-        await Video.findOneAndRemove({_id:id});
+        const video = await Video.findById(id);
+        if(video.creator.toString() !== req.user.id) {
+            throw Error();
+        } else {
+            await Video.findOneAndRemove({_id:id});
+        }
     } catch(error) {
         console.log(error);
     }
     res.redirect(routes.home);
 }
+
+export const postRegisterView = async (req, res) => {
+    const {
+        params: {id}
+    } = req;
+    try {
+        const video = await Video.findById(id);
+        video.views += 1;
+        video.save();
+        res.status(200);
+    } catch(error) {
+        res.status(400);
+        res.end();
+    } finally {
+        res.end();
+    }
+}
+
+export const postAddComment = async (req, res) => {
+    const {
+        params: {id},
+        body: {comment},
+        user
+    } = req;
+    try {
+        const video = await Video.findById(id);
+        const newComment = await Comment.create({
+            text: comment,
+            creator: user.id
+        });
+        video.comments.push(newComment._id);
+        video.save();
+    }catch(error) {
+        res.status(400);
+    } finally {
+        res.end();
+    }
+};
